refactor(fe): 배틀페이지에서 목 데이터 세팅 처리 분리
useTutorialBattleSetup.ts 컴포넌트로 스토어 초기화, 목 소켓 연결, 정리 책임 분리
